+++ datastore/couchDB/utilities.go
@@ -4,6 +4,7 @@ import (
 	"encoding/json"
 	"fmt"
 	"net/url"
+	"strconv"
 	"strings"
 
 	ds "github.com/accedian/adh-gather/datastore"
@@ -12,7 +13,14 @@ import (
 	couchdb "github.com/leesper/couchdb-golang"
 )
 
-const defaultQueryResultsLimit = 1000
+const (
+	defaultQueryResultsLimit = 1000
+
+	startKeyQueryParamStr    = "startkey"
+	limitQueryParamStr       = "limit"
+	includeDocsQueryParamStr = "include_docs"
+	descendingQueryParamStr  = "descending"
+)
 
 // ConvertDataToCouchDbSupportedModel - Turns any object into a CouchDB ready entry
 // that can be stored. Changes the provided object into a map[string]interface{} generic
@@ -130,6 +138,9 @@ func getByDocIDWithQueryParams(docID string, dataTypeStrForLogging string, db *c
 	if queryParams == nil {
 		queryParams = new(url.Values)
 	}
+
+	logger.Log.Warningf("Issuing query for %s with params %v", docID, queryParams)
+
 	fetchedData, err := db.Get(docID, *queryParams)
 	if err != nil {
 		logger.Log.Debugf("Error retrieving %s %s: %s", dataTypeStrForLogging, docID, err.Error())
@@ -556,4 +567,19 @@ func getAllInIDListFromCouchAndFlatten(dbName string, idList []string, dataType 
 	// Marshal the response from the datastore to bytes so that it
 	// can be Marshalled back to the proper type.
 	return convertCouchDataArrayToFlattenedArray(fetchedList, dataContainer, loggingStr)
+}
+
+func generatePaginationQueryParams(startKey string, limit int64, includeDocs bool, descending bool) url.Values {
+	params := url.Values{}
+
+	// pges offset in couchDB terms is index value, or number of pages * number of records per page
+	if len(startKey) != 0 {
+		params.Add(startKeyQueryParamStr, fmt.Sprintf(`"%s"`, startKey))
+	}
+
+	params.Add(limitQueryParamStr, strconv.FormatInt(limit, 10))
+	params.Add(includeDocsQueryParamStr, strconv.FormatBool(includeDocs))
+	params.Add(descendingQueryParamStr, strconv.FormatBool(descending))
+
+	return params
 }