+++ models/common/common.go
@@ -30,3 +30,10 @@ type BulkOperationResult struct {
 	ERROR  string `json:"error"`
 	REASON string `json:"reason"`
 }
+
+// PaginationOffsets - struct to hold offsets for pagination related queries
+type PaginationOffsets struct {
+	Self string
+	Prev string
+	Next string
+}